+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All major and minor version changes will be documented in this file. Details of
 patch-level version changes can be found in [commit messages](../../commits/master).
 
+## 2023.0.2 - 2023/06/27
+
+- Add `Level.CRIT`
+- Misc fixes to make plugins more resilient
+- Test plugin utility functions
+
 ## 2023.0.1 - 2023/06/27
 
 - Fix `csv` writer